requirements.txt
aiodns==1.1.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
ccxt==1.18.1270
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
idna==2.8
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
python-dateutil==2.8.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
yarl==1.1.0
